requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==1.0.4
gotrue==1.0.3
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.11.7
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.24.1
email-validator==2.1.0.post1
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
redis==5.0.1
razorpay==1.3.0
requests==2.31.0
pytz==2023.3
